requirements.txt
pypdf>=5.4.0
python-docx>=1.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
beautifulsoup4>=4.12.0
PyMuPDF>=1.24.0
easyocr>=1.7.0
